+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # Version 2.0.0 in progress
 
 * Make `(defexpect foo)` and `(defexpect foo (bar))` match the behavior of `deftest`, without wrapping the body in `(expect ,,,)`. This is potentially breaking insofar as `(defexpect foo (produces-falsey))` would have been a failing test in 1.x but now silently just runs `(produces-falsey)` in the same way that `(deftest foo (produces-falsey))` does.
-* Add `before`/`after`/`around` for each/once fixtures, as an alternative for having to require `clojure.test` just for `use-fixtures`. This is an experimental feature and is not yet documented!
-* Bring several test-running functions from `clojure.test` in with "expect" names, for convenience in dev/test so users don't need to require `clojure.test` as well.
+* Bring several test-running functions from `clojure.test` in, for convenience in dev/test so users don't need to require `clojure.test` as well.
 * Add various macro-like constructs back into the documentation to improve the documentation (`in`, `from-each`, `more-of`, `more->`, `more` are really only syntactic constructs inside `expect`).
 
 # Version 1.2.1 -- 2019-12-09